Moving to a pear upgrader that doesn't prompt
In using this chef recipe in a non vagrant environment we get a prompt.  As a
result you have to hit enter.  Doing a bit a research I came up with this.

http://old.nabble.com/go-pear.phar---command-line-options--td17172511.html

We moved towards it. I tested it and it seems to have the same effect.  Failing
this we need to start detecting which version of PEAR and distribution we are
using an start to use the native PEAR updating tool.